Fixes a bad counting bug
When the names are split into parts to ascertain the array structure, you get
some blank ones. The previous method of cleaning out blank elements was no good
because it made it difficult to detect the end of the array. this way is better
as it filters out the bad items before we get to processing.